requirements.txt
fastapi
uvicorn
python-multipart
langgraph
langchain
langchain-openai
llama-index
llama-index-embeddings-huggingface
deepgram-sdk
python-dotenv
streamlit>=1.40.0
sqlalchemy
psycopg2-binary
boto3
pinecone-client
llama-index-vector-stores-pinecone

# ML core
torch>=2.0.0
transformers>=4.35.0
sentence-transformers>=2.3.0
numpy>=1.24.0
pandas>=2.0.0

# Audio processing
librosa>=0.10.0
soundfile>=0.12.0

# Training utilities
scikit-learn>=1.3.0
scipy>=1.11.0
tensorboard>=2.16.0
matplotlib>=3.7.0
seaborn>=0.12.0

# RL
gymnasium>=0.29.1
stable-baselines3>=2.3.0

# Retrieval
rank-bm25>=0.2.2

# Explainability (Phase 5)
shap>=0.45.0

# LLM clients
groq>=0.4.0
